requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0
cloudscraper>=1.2.0
beautifulsoup4>=4.12.0
flask>=3.0.0
openai>=1.0.0
mcp>=0.9.0
streamlit>=1.28.0
streamlit>=1.28.0

